Make touch input activate(it still dosent work, but it does activate :) ) without pushing the up button
Also fix some reported bugs.

+++ src/68328Functions.c
@@ -102,7 +102,7 @@ void patchTo68328(){
 
    patchOpcode(OPCODE_BGND, m68k_op_bgnd, 16/*dont know how many cycles, average opcode*/);
    
-   for(uint32_t currentOpcode = OPCODE_CPU32_START; currentOpcode <= OPCODE_CPU32_END; currentOpcode++)
+   for(uint16_t currentOpcode = OPCODE_CPU32_START; currentOpcode <= OPCODE_CPU32_END; currentOpcode++)
       patchOpcode(currentOpcode, m68k_op_cpu32_dispatch, 91/*dont know how many cycles, most expensive opcode*/);
 
    //68328 may actually use some 68020 opcodes, those will be patched in if and when Palm OS attempts to call one
